requirements.txt
chardet
jinja2
pyyaml